fixed padding, added login page
There is the main issue of a <div> not being able to auto adjust to screen size. The home page on my computer has 2 scroll bars when i do w-screen h-screen, I want to be able to have the <div> fit the users screen size and not need scroll bars. Not sure yet how to do this