image for login(placeholder)
quiann drew this, looks really nice. I'm gonna give her the hex code for the header color, bc the pickup tool didnt grab the right shade, and then it'll blend into the header. Besides that looks nice. I can figure out the dimensions of the image and stretch it later when it's done. The actual login screen will be really simple: a rounded box with a email/password enter bar and a button.